execution/requirements_thumbnail_gen.txt
# Requirements for generate_thumbnail.py (simplified local thumbnail generation)
#
# Install with: pip install -r requirements_thumbnail_gen.txt
#
# GPU Requirements:
#   - NVIDIA: 8GB+ VRAM (RTX 3060, 3070, 4060, etc.)
#   - Apple Silicon: Use --device mps
#   - CPU: Works but very slow (~5-10 min per image)

torch>=2.0.0
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.25.0
pillow>=10.0.0

# Optional: extra memory savings (recommended for 8GB cards)
# xformers>=0.0.23
